+++ delta_backend/tests/utils_for_converter_tests.py
@@ -1,5 +1,17 @@
 from decimal import Decimal
+from helpers.mappings import EventName, Operation
+from sample_data.test_resource_data import get_test_data_resource
 import json
+from typing import List
+
+
+class RecordConfig:
+    def __init__(self, event_name, operation, imms_id, expected_action_flag=None, supplier="EMIS"):
+        self.event_name = event_name
+        self.operation = operation
+        self.supplier = supplier
+        self.imms_id = imms_id
+        self.expected_action_flag = expected_action_flag
 
 
 class ValuesForTests:
@@ -118,43 +130,69 @@ class ValuesForTests:
     json_value_for_test = json.dumps(json_data)
 
     @staticmethod
-    def get_event(event_name="INSERT", operation="CREATE", supplier="EMIS"):
-        if operation != "REMOVE":
-            return {
-                "Records": [
-                    {
-                        "eventName": event_name,
-                        "dynamodb": {
-                            "ApproximateCreationDateTime": 1690896000,
-                            "NewImage": {
-                                "PK": {"S": "covid#12345"},
-                                "PatientSK": {"S": "COVID19#ca8ba2c6-2383-4465-b456-c1174c21cf31"},
-                                "IdentifierPK": {"S": "system#1"},
-                                "Operation": {"S": operation},
-                                "SupplierSystem": {"S": supplier},
-                                "Resource": {"S": ValuesForTests.json_value_for_test},
-                            },
-                        },
+    def get_multi_record_event(records_config: List[RecordConfig]):
+        records = []
+        for config in records_config:
+            # Extract values from the config dictionary
+            imms_id = config.imms_id
+            event_name = config.event_name
+            operation = config.operation
+            supplier = config.supplier
+
+            # Generate record using the provided configuration
+            records.append(
+                ValuesForTests.get_test_event_record(
+                    imms_id=imms_id,
+                    event_name=event_name,
+                    operation=operation,
+                    supplier=supplier,
+                )
+            )
+        return {"Records": records}
+
+    @staticmethod
+    def get_event(event_name=EventName.CREATE, operation=Operation.CREATE, supplier="EMIS", imms_id="12345"):
+        """Create test event for the handler function."""
+        return {
+            "Records": [
+                ValuesForTests.get_event_record(imms_id, event_name, operation, supplier)
+            ]
+        }
+
+    @staticmethod
+    def get_event_record(imms_id, event_name, operation, supplier="EMIS"):
+        pk = f"covid#{imms_id}"
+        if operation != Operation.DELETE_PHYSICAL:
+            return{
+                "eventName": event_name,
+                "dynamodb": {
+                    "ApproximateCreationDateTime": 1690896000,
+                    "NewImage": {
+                        "PK": {"S": pk},
+                        "PatientSK": {"S": pk},
+                        "IdentifierPK": {"S": "system#1"},
+                        "Operation": {"S": operation},
+                        "SupplierSystem": {"S": supplier},
+                        "Resource": {
+                            "S": json.dumps(get_test_data_resource()),
+                        }
                     }
-                ]
+                }
             }
         else:
             return {
-                "Records": [
-                    {
-                        "eventName": "REMOVE",
-                        "dynamodb": {
-                            "ApproximateCreationDateTime": 1690896000,
-                            "Keys": {
-                                "PK": {"S": "covid#12345"},
-                                "PatientSK": {"S": "covid#12345"},
-                                "SupplierSystem": {"S": "EMIS"},
-                                "Resource": {"S": ValuesForTests.json_value_for_test},
-                                "PatientSK": {"S": "COVID19#ca8ba2c6-2383-4465-b456-c1174c21cf31"},
-                            },
-                        },
+                "eventName": event_name,
+                "dynamodb": {
+                    "ApproximateCreationDateTime": 1690896000,
+                    "Keys": {
+                        "PK": {"S": pk},
+                        "PatientSK": {"S": pk},
+                        "SupplierSystem": {"S": supplier},
+                        "Resource": {
+                            "S": json.dumps(get_test_data_resource()),
+                        }
                     }
-                ]
+                }
             }
 
     expected_static_values = {
